Revert back to .c and .h implementation
